rework capnp::Error to include updated RPC Exception fields
With the updated RPC schema, we've access to `details`, allowing
servers to send information back to clients on errors. This is useful
more application-level details, like validation errors.

This required a deep refactoring of the `capnp::Error` type. Previously
it was a pair of public pairs (kind, reason). To support efficiently
details and in the future other fields, we restructured it be more
similar to `io::Error`, storing kind and a boxed "inner" struct, which
contains the more heavyweight data.